+++ app/code/community/Jtl/Connector/sql/jtl_connector_setup/upgrade-1.3.0.0-1.4.4.0.php
@@ -7,6 +7,5 @@ $sql = "
       CHANGE `foreign_key` `foreign_key` INT(11) DEFAULT 0  NOT NULL,
       ADD COLUMN `endpoint_id` INT(11) DEFAULT 0  NOT NULL AFTER `foreign_key`;
     ";
-error_log($sql);
 $installer->run($sql);
 $installer->endSetup();