+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-1.4.4.0 (unreleased)
+1.4.5.0 (unreleased)
 --------------------
+
+1.4.4.0
+-------
+- Limit all operations to simple and configurable products
+- Support discount codes
 
 1.4.3.0
 -------